pyproject.toml
[project]
name = "calendar-event-converter"
version = "0.1.0"
description = "Convert unstructured text and documents into CalDAV calendar events"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "icalendar>=5.0.11",
    "openai>=1.10.0",
    "pyyaml>=6.0.1",
    "hypothesis>=6.98.0",
    "pytest>=8.0.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
